Find recently hired candidates in Jobvite.
We have to jump through some hoops to escape the query parameters for the
Jobvite API: Ruby's CGI.escape method encodes spaces as "+" (e.g.
"Offer+Accepted"), but using a "+" character causes an exception in
Jobvite's API. CGI.escape will percent-encode "+" characters in the input,
so we can safely assume that any "+" characters in the output are encoded
spaces, and safely replace them with "%20".

This commit introduces a `NamelyImporter` object, which will be responsible
for sending the candidate information from Jobvite into Namely. Currently,
this object is just a stub.

https://trello.com/c/rBzLZ0wQ