Users can connect to NetSuite
This commit implements a basic NetSuite client and follows a similar
workflow to other connections for setting up a NetSuite account.

The NetSuite connection is a little different, because it doesn't store
NetSuite credentials. It makes a request to Cloud Elements with the
user's credentials, and then stores an authorization token.

No import or export functionality is added for NetSuite. This commit
only adds the UI and backend for establishing a connection. It will
serve as a starting point for future commits.

The view doesn't show the NetSuite connection UI unless a Cloud Elements
API is configured. This means that development/test environments can see
the NetSuite UI, but staging and production will not until we're ready.

https://trello.com/c/HeQUdynL