First pass at AttributeMapper and FieldMapping in ActiveRecord
* AttributeMapper belongs to a User, primarily so if a User goes away,
  the AttributeMapper will, too
* AttributeMappers will be created per connection. AttributeMappers
* have_a
  connection, Connections belong_to an AttributeMapper
  * This is in order to avoid needing a polymorphic relationship on
    AttributeMapper
* NB: Not yet wired into NetSuite::Export NB: Does not yet handle fields
* that need any sort of additional handling

Following a recommendation from Mike Burns to not create this implicitly
by calling `#attribute_mapper`, particularly when a GET request uses
NetSuite::Connection, it may have database side-effects, which violates
the Principle of Least Surprise.

We're not thrilled with having it in an `after_create` here, but we like
it better than it being implicit.

* For: https://trello.com/c/JsdID6YF/