Generate better instance names on Cloud Elements
Previously, we had many conflicting instance names because they were
keyed only on the account id. This was confusing in development, but
potentially dangerous as we rolled out to other environments where the
instance names would conflict yet again.

This change extract the responsibility of assembling the parameters for
creating an instance to the `NetSuite::Instance` object. Names will now
have both the Rails environment and a timestamp added to the name to
help make finding the correct cloud elements instance easier.