Clean up dashboard display
* Move edit actions into a readable list
* Make all button styles consistent
* Move disconnect button into header to mirror "connect"
* Hide webhook URLs behind "Show Webhook URL" button.

Having a random URL hanging around the UI for iCIMS and Greenhouse
doesn't make a lot of sense and is also not useful to the user most of
the time. By tucking it behind a button, we can both better label it and
keep it from cluttering the UI in the majority of cases where no one
needs to see it.