Update to latest Rails 4.2
* Configure active_job
* Fix deprecations

`deliver` has been deprecated and replaced with `deliver_now` and
`deliver_later` (via ActiveJob). For now, I used `deliver_now` as the
fastest way to get the deprecations fixed while maintaining a green test
suite. The mailers are run in response to background jobs already, so I
think this is preferable.