Add activity feed for NetSuite syncs
This gives users a web-accessible view of the sync activity happening
with their net suite integration. The data is similar to what is emailed
to them.

The view currently does not distinguish between successful or
unsuccessful profile events, but we have a follow-on task that will take
care of that.

https://trello.com/c/WsHL9kNH