Instantiate SyncJob with a connection
This previously took two arguments, both of which are accessible via the
connection. We had convenient access to the connection in all places
where we were instantiating Sync Jobs.

Ordinarily, persisting ActiveRecord objects for a background job is not
a good idea, but by converting to ActiveJob/GlobalId, that issue is
solved.