Send all unauthorized emails from notifier
Previously, the unauthorized email was being sent via model method on
`Installation`. The `UnauthorizedNotifier` used this as did a couple of
places in other code.

I removed the ActiveRecord model methods concerned with mail delivery
and opted to send the mail directly via the `UnauthorizedNotifier` just
as we do with `SyncJob`. Then all places that were sending the mail
directly, without the notifier, were updated to use the notifier.

This centralizes the responsibility for tracking unauthorized sync
summaries and ensures we get both an email and a summary object created.