Fix NetSuite employee creation
The Namely API actually return empty string rather than null for new
employees.

This has always been the case, but a recent refactoring changed it to
check for null (which we were using in our fixtures) rather than empty
string.

This changes the fixtures to be more realistic, so that we'll catch
future regressions, and fixes the NetSuite client to create a new
employee when it receives an empty string ID.