Handle more Errors from NetSuite
We occasionally get errors back from NetSuite other than 400 (Bad
Request) or 401 (Unauthorized). When this happense, we're crashing the
entire job.

By handling `RestClient::Exception`, we can handle *all*  unexpected
HTTP response codes, allowing the job to continue running to completion
and record a proper `SyncSummary`.

We will now log and report all API errors to Raygun. If this gets noisy,
we can adjust.

Separate raising from handling, log in Raygun

Let's log all API errors that occur in the export to Raygun so we have
good visibility into the errors.

In doing so, let's also separate the code that raises exceptions from
the code that logs/handles them, so the decision to log happens with the
decision to handle.