Report DelayedJob exceptions to Raygun
Raygun doesn't ship with support for DelayedJob out of the box, so we
have to use DelayedJob's plugin system to catch and report these. I
based this plugin off of this: http://blog.salsify.com/engineering/delayed-jobs-callbacks-and-hooks-in-rails

I could not conceive of a good way to test this in an automated fashion,
so I tested it in development by setting SyncJob to raise from `perform`
and making sure the `rescue` block in the plugin was executing and that
DelayedJob still requeued the job for a retry.