Improve reporting of API errors
Form a reporting perspective, we want to capture as much context from
the original exception as possible. By logging the original exception,
we'll easily be able to differentiate the different subclasses of
`RestClient::Exception` We are only raising a different exception in
order to isolate our application from the RestClient dependency.

Finally, when we create the NetSuite::ApiError, we want to be sure to
include the pre-existing backtrace so we can get good stacktraces
wherever this exception ends up peeking out again.