Allow failed profiles to be retried
An installation may have hundreds or even thousands of namely profiles
with only a small number of failures. In these cases, it's reasonable
for users to want to quickly retry only the failed profiles once they
think they have made sufficient corrections.

To facilitate, this change introduces a `RetryJob` which will sync only
the failed profiles associated with a sync summary. Syncing was built to
sync all namely profiles over a given connection, so we needed a way to
tell the sync operation to only sync a subset of profiles. An optional
named argument of `profiles` was added to the net suite sync.

A `Notifer` class was also extracted to handle the common concern of
notifications between the `SyncJob` and `RetryJob`.