Provide Greenhouse and Jobvite documentation
I understand these integrations far less than NetSuite, so their
documentation is not nearly as detailed. Jobvite is incomplete as I
haven't yet sussed out all of the details of how the normalizer and
attribute mappers work in that case.

I included a PDF of the applicable Jobvite API documentation because it
is not accessible online.